[Bronze V] Title: 이중 반복문, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Bronze/34849. 이중 반복문/이중 반복문.cc
@@ -4,6 +4,6 @@ using namespace std;
 int main(void) {
     long n;
     cin>>n;
-    cout<<((double)n*n/pow(10,8)>1?"Time limit exceeded":"Accepted");
+    cout<<(n*n/pow(10,8)>1?"Time limit exceeded":"Accepted");
     return 0;
 }
